rafactor: useProgressAnimation 훅 분리
