+++ app/application/_components/ProgressBar.tsx
@@ -1,8 +1,8 @@
-import { useEffect, useRef } from "react";
 import { Animated, View } from "react-native";
 
 import Flex from "@/src/components/layout/Flex";
 import TextSize from "@/src/components/ui/TextSize";
+import { useProgressAnimation } from "@/src/hooks/application/useProgressAnimation";
 import { COLORS } from "@/src/utils/constants/colors";
 
 interface ProgressBarProps {
@@ -17,25 +17,10 @@ export default function ProgressBar({
   stepTitle,
 }: ProgressBarProps) {
   const progressPercent = (currentStep / totalSteps) * 100;
-
   const currentStepText = `${currentStep}/${totalSteps}`;
   const titleText = `${currentStep}단계: ${stepTitle}`;
 
-  const loadValue = useRef(new Animated.Value(0)).current;
-
-  useEffect(() => {
-    Animated.timing(loadValue, {
-      toValue: progressPercent,
-      duration: 500,
-      useNativeDriver: false,
-    }).start();
-  }, [currentStep]);
-
-  const loadWidth = loadValue.interpolate({
-    inputRange: [0, 100],
-    outputRange: ["0%", "100%"],
-    extrapolate: "clamp",
-  });
+  const { loadWidth } = useProgressAnimation(progressPercent);
 
   return (
     <View className="p-3">
@@ -47,7 +32,7 @@ export default function ProgressBar({
           content={currentStepText}
         />
       </Flex>
-      
+
       <View className="h-2" />
 
       <View className="h-2 bg-gray-border rounded-full">
